Move dockecompose to pkg
