+++ newrelic-integration-e2e/pkg/dockercompose/docker.go
